# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Authentication is handled using JWT tokens. The tokens are set as access-token and refresh-token cookies.\nFor information regarding which role is authorized for which endpoint, see the http handler https://github.com/Zibbp/ganymede/blob/main/internal/transport/http/handler.go.",
        "title": "Ganymede API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1",
    "paths": {
...
        "/playback": {
            "get": {
                "security": [
                    {
                        "ApiKeyCookieAuth": []
                    }
                ],
                "description": "Get all playback progress",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Playback"
                ],
                "summary": "Get all progress",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Playback"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/playback/last": {
            "get": {
                "security": [
                    {
                        "ApiKeyCookieAuth": []
                    }
                ],
                "description": "Get last playback entries for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Playback"
                ],
                "summary": "Get last playbacks",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/playback.GetPlaybackResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/playback/progress": {
            "post": {
                "security": [
                    {
                        "ApiKeyCookieAuth": []
                    }
                ],
                "description": "Update playback progress",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Playback"
                ],
                "summary": "Update progress",
                "parameters": [
                    {
                        "description": "progress",
                        "name": "progress",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdateProgressRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/playback/status": {
            "post": {
                "security": [
                    {
                        "ApiKeyCookieAuth": []
                    }
                ],
                "description": "Update playback status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Playback"
                ],
                "summary": "Update status",
                "parameters": [
                    {
                        "description": "status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/playback/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyCookieAuth": []
                    }
                ],
                "description": "Get playback progress",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Playback"
                ],
                "summary": "Get progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "vod id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Playback"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyCookieAuth": []
                    }
                ],
                "description": "Delete playback progress",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Playback"
                ],
                "summary": "Delete progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "vod id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ent.Channel": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "display_name": {
                    "description": "DisplayName holds the value of the \"display_name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ChannelQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.ChannelEdges"
                        }
                    ]
                },
                "ext_id": {
                    "description": "The external ID of the channel.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "string"
                },
                "image_path": {
                    "description": "ImagePath holds the value of the \"image_path\" field.",
                    "type": "string"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "retention": {
                    "description": "Retention holds the value of the \"retention\" field.",
                    "type": "boolean"
                },
                "retention_days": {
                    "description": "RetentionDays holds the value of the \"retention_days\" field.",
                    "type": "integer"
                },
                "storage_size_bytes": {
                    "description": "Total storage size in bytes for the channel's videos.",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Chapter": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ChapterQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.ChapterEdges"
                        }
                    ]
                },
                "end": {
                    "description": "End holds the value of the \"end\" field.",
                    "type": "integer"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "string"
                },
                "start": {
                    "description": "Start holds the value of the \"start\" field.",
                    "type": "integer"
                },
                "title": {
                    "description": "Title holds the value of the \"title\" field.",
                    "type": "string"
                },
                "type": {
                    "description": "Type holds the value of the \"type\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Playback": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "string"
                },
                "status": {
                    "description": "Status holds the value of the \"status\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.PlaybackStatus"
                        }
                    ]
                },
                "time": {
                    "description": "Time holds the value of the \"time\" field.",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                },
                "user_id": {
                    "description": "UserID holds the value of the \"user_id\" field.",
                    "type": "string"
                },
                "vod_id": {
                    "description": "VodID holds the value of the \"vod_id\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Vod": {
            "type": "object",
            "properties": {
                "caption_path": {
                    "description": "CaptionPath holds the value of the \"caption_path\" field.",
                    "type": "string"
                },
                "chat_path": {
                    "description": "ChatPath holds the value of the \"chat_path\" field.",
                    "type": "string"
                },
                "chat_video_path": {
                    "description": "ChatVideoPath holds the value of the \"chat_video_path\" field.",
                    "type": "string"
                },
                "clip_ext_vod_id": {
                    "description": "The external VOD ID of a clip. This is only populated if the clip is linked to a video.",
                    "type": "string"
                },
                "clip_vod_offset": {
                    "description": "The offset in seconds to where the clip starts in the VOD. This is only populdated if the video is a clip.",
                    "type": "integer"
                },
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "duration": {
                    "description": "Duration holds the value of the \"duration\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the VodQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.VodEdges"
                        }
                    ]
                },
                "ext_id": {
                    "description": "The ID of the video on the external platform.",
                    "type": "string"
                },
                "ext_stream_id": {
                    "description": "The ID of the stream on the external platform, if applicable.",
                    "type": "string"
                },
                "file_name": {
                    "description": "FileName holds the value of the \"file_name\" field.",
                    "type": "string"
                },
                "folder_name": {
                    "description": "FolderName holds the value of the \"folder_name\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "string"
                },
                "info_path": {
                    "description": "InfoPath holds the value of the \"info_path\" field.",
                    "type": "string"
                },
                "live_chat_convert_path": {
                    "description": "Path to the converted live chat file",
                    "type": "string"
                },
                "live_chat_path": {
                    "description": "Path to the raw live chat file",
                    "type": "string"
                },
                "locked": {
                    "description": "Locked holds the value of the \"locked\" field.",
                    "type": "boolean"
                },
                "notes": {
                    "description": "User notes about why the VOD was kept.",
                    "type": "string"
                },
                "platform": {
                    "description": "The platform the VOD is from, takes an enum.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.VideoPlatform"
                        }
                    ]
                },
                "processing": {
                    "description": "Whether the VOD is currently processing.",
                    "type": "boolean"
                },
                "resolution": {
                    "description": "Resolution holds the value of the \"resolution\" field.",
                    "type": "string"
                },
                "sprite_thumbnails_columns": {
                    "description": "SpriteThumbnailsColumns holds the value of the \"sprite_thumbnails_columns\" field.",
                    "type": "integer"
                },
                "sprite_thumbnails_enabled": {
                    "description": "SpriteThumbnailsEnabled holds the value of the \"sprite_thumbnails_enabled\" field.",
                    "type": "boolean"
                },
                "sprite_thumbnails_height": {
                    "description": "SpriteThumbnailsHeight holds the value of the \"sprite_thumbnails_height\" field.",
                    "type": "integer"
                },
                "sprite_thumbnails_images": {
                    "description": "SpriteThumbnailsImages holds the value of the \"sprite_thumbnails_images\" field.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sprite_thumbnails_interval": {
                    "description": "SpriteThumbnailsInterval holds the value of the \"sprite_thumbnails_interval\" field.",
                    "type": "integer"
                },
                "sprite_thumbnails_rows": {
                    "description": "SpriteThumbnailsRows holds the value of the \"sprite_thumbnails_rows\" field.",
                    "type": "integer"
                },
                "sprite_thumbnails_width": {
                    "description": "SpriteThumbnailsWidth holds the value of the \"sprite_thumbnails_width\" field.",
                    "type": "integer"
                },
                "storage_size_bytes": {
                    "description": "The size of the VOD in bytes.",
                    "type": "integer"
                },
                "streamed_at": {
                    "description": "The time the VOD was streamed.",
                    "type": "string"
                },
                "thumbnail_path": {
                    "description": "ThumbnailPath holds the value of the \"thumbnail_path\" field.",
                    "type": "string"
                },
                "title": {
                    "description": "Title holds the value of the \"title\" field.",
                    "type": "string"
                },
                "tmp_chat_download_path": {
                    "description": "The path where the chat is downloaded to",
                    "type": "string"
                },
                "tmp_chat_render_path": {
                    "description": "The path where the rendered chat is",
                    "type": "string"
                },
                "tmp_live_chat_convert_path": {
                    "description": "The path where the converted chat is",
                    "type": "string"
                },
                "tmp_live_chat_download_path": {
                    "description": "The path where the converted chat is",
                    "type": "string"
                },
                "tmp_video_convert_path": {
                    "description": "The path where the converted video is",
                    "type": "string"
                },
                "tmp_video_download_path": {
                    "description": "The path where the video is downloaded to",
                    "type": "string"
                },
                "tmp_video_hls_path": {
                    "description": "The path where the temporary video hls files are",
                    "type": "string"
                },
                "type": {
                    "description": "The type of VOD, takes an enum.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.VodType"
                        }
                    ]
                },
                "updated_at": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                },
                "video_hls_path": {
                    "description": "The path where the video hls files are",
                    "type": "string"
                },
                "video_path": {
                    "description": "VideoPath holds the value of the \"video_path\" field.",
                    "type": "string"
                },
                "web_thumbnail_path": {
                    "description": "WebThumbnailPath holds the value of the \"web_thumbnail_path\" field.",
                    "type": "string"
                }
            }
        },
        "ent.VodEdges": {
            "type": "object",
            "properties": {
                "channel": {
                    "description": "Channel holds the value of the channel edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.Channel"
                        }
                    ]
                },
                "chapters": {
                    "description": "Chapters holds the value of the chapters edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Chapter"
                    }
                },
                "multistream_info": {
                    "description": "MultistreamInfo holds the value of the multistream_info edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.MultistreamInfo"
                    }
                },
                "muted_segments": {
                    "description": "MutedSegments holds the value of the muted_segments edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.MutedSegment"
                    }
                },
                "playlists": {
                    "description": "Playlists holds the value of the playlists edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Playlist"
                    }
                },
                "queue": {
                    "description": "Queue holds the value of the queue edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.Queue"
                        }
                    ]
                }
            }
        },
...
        "http.UpdateProgressRequest": {
            "type": "object",
            "required": [
                "time",
                "vod_id"
            ],
            "properties": {
                "time": {
                    "type": "integer"
                },
                "vod_id": {
                    "type": "string"
                }
            }
        },
...
        "http.UpdateStatusRequest": {
            "type": "object",
            "required": [
                "status",
                "vod_id"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "in_progress",
                        "finished"
                    ]
                },
                "vod_id": {
                    "type": "string"
                }
            }
        },
...
        "playback.GetPlayback": {
            "type": "object",
            "properties": {
                "playback": {
                    "$ref": "#/definitions/ent.Playback"
                },
                "vod": {
                    "$ref": "#/definitions/ent.Vod"
                }
            }
        },
        "playback.GetPlaybackResp": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/playback.GetPlayback"
                    }
                },
                "playback": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Playback"
                    }
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "utils.PlaybackStatus": {
            "type": "string",
            "enum": [
                "in_progress",
                "finished"
            ],
            "x-enum-varnames": [
                "InProgress",
                "Finished"
            ]
        },
...
        "utils.VideoPlatform": {
            "type": "string",
            "enum": [
                "twitch",
                "youtube"
            ],
            "x-enum-varnames": [
                "PlatformTwitch",
                "PlatformYoutube"
            ]
        },
...
        "utils.VodType": {
            "type": "string",
            "enum": [
                "archive",
                "live",
                "highlight",
                "upload",
                "clip"
            ],
            "x-enum-varnames": [
                "Archive",
                "Live",
                "Highlight",
                "Upload",
                "Clip"
            ]
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Authorization: Bearer \u003capi_key\u003e. Mint a key via the admin UI.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ApiKeyCookieAuth": {
            "type": "apiKey",
            "name": "access-token",
            "in": "cookie"
        }
    }
}